ai-service/requirements.txt
fastapi
uvicorn
requests
pydantic
sentence-transformers
faiss-cpu
numpy
